раздробил основной фаил на подфайлы
